refactor: simplify review cleanup
- Extract makeThrowingScope to shared helper in ports/__testHelpers__ —
  was duplicated identically in three test files
- Drop dead optional chaining in name-matches.ts: use getParentOrThrow()
  since Identifier nodes always have a parent in ts-morph
- Pre-build local-name → import-meta Map in collectTransitiveImports,
  removing the O(N×M) nested import-declaration loop
- Remove redundant WHAT JSDoc from findNonExportedDeclaration and
  resolveDeclarationStatement
