+++ src/ts-engine/transitive-imports.ts
@@ -15,6 +15,24 @@ export function collectTransitiveImports(
     { namedImports: Array<{ name: string; alias?: string }>; resolvedAbsPath: string }
   >();
 
+  // Pre-build a map from local name → import metadata so the identifier loop
+  // is O(identifiers) rather than O(identifiers × imports).
+  type ImportMeta = { originalName: string; aliasText: string | undefined; importedPath: string };
+  const localNameToImport = new Map<string, ImportMeta>();
+  for (const importDecl of srcSF.getImportDeclarations()) {
+    const importedSF = importDecl.getModuleSpecifierSourceFile();
+    if (!importedSF) continue;
+    const importedPath = importedSF.getFilePath() as string;
+    for (const ni of importDecl.getNamedImports()) {
+      const localName = ni.getAliasNode()?.getText() ?? ni.getName();
+      localNameToImport.set(localName, {
+        originalName: ni.getName(),
+        aliasText: ni.getAliasNode()?.getText(),
+        importedPath,
+      });
+    }
+  }
+
   for (const identifier of declStmt.getDescendantsOfKind(SyntaxKind.Identifier)) {
     const sym = identifier.getSymbol();
     if (!sym) continue;
@@ -34,39 +52,22 @@ export function collectTransitiveImports(
     )
       continue;
 
-    for (const importDecl of srcSF.getImportDeclarations()) {
-      const namedImports = importDecl.getNamedImports();
-      const matchingNamed = namedImports.find((ni) => {
-        const localName = ni.getAliasNode()?.getText() ?? ni.getName();
-        return localName === identifier.getText();
-      });
-      if (!matchingNamed) continue;
+    const importMeta = localNameToImport.get(identifier.getText());
+    if (!importMeta || importMeta.importedPath !== resolvedPath) continue;
 
-      const importedSF = importDecl.getModuleSpecifierSourceFile();
-      if (!importedSF) continue;
-      const importedPath = importedSF.getFilePath() as string;
-      if (importedPath !== resolvedPath) continue;
+    const { originalName, aliasText, importedPath } = importMeta;
+    const namedImportEntry =
+      aliasText && aliasText !== originalName
+        ? { name: originalName, alias: aliasText }
+        : { name: originalName };
 
-      const originalName = matchingNamed.getName();
-      const aliasNode = matchingNamed.getAliasNode();
-      const aliasText = aliasNode?.getText();
-      const namedImportEntry =
-        aliasText && aliasText !== originalName
-          ? { name: originalName, alias: aliasText }
-          : { name: originalName };
-
-      if (!result.has(importedPath)) {
-        result.set(importedPath, {
-          namedImports: [namedImportEntry],
-          resolvedAbsPath: importedPath,
-        });
-      } else {
-        const existing = result.get(importedPath);
-        if (existing && !existing.namedImports.some((ni) => ni.name === namedImportEntry.name)) {
-          existing.namedImports.push(namedImportEntry);
-        }
+    if (!result.has(importedPath)) {
+      result.set(importedPath, { namedImports: [namedImportEntry], resolvedAbsPath: importedPath });
+    } else {
+      const existing = result.get(importedPath);
+      if (existing && !existing.namedImports.some((ni) => ni.name === namedImportEntry.name)) {
+        existing.namedImports.push(namedImportEntry);
       }
-      break;
     }
   }
